pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qsml"
version = "0.1.0"
description = "Stochastic Volatility & Options: Heston Calibration with ML Surrogate"
authors = [{name = "Finance Quant", email = "quant@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "numba>=0.57.0",
    "matplotlib>=3.7.0",
    "plotly>=5.14.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "torchmetrics>=0.11.0",
    "optuna>=3.2.0",
    "statsmodels>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "seaborn>=0.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"